Fix landing page layout
The landing page layout was restricting children elements to 1200px max-width. This meant that background-colors could not expand to the full width of the page. Allowing each element to take up 100% width but have its own styles that restrict it to a max-width/centered horizontally on the page fixes this.