Explicitly set "domain" on session cookie
Cookies are set to the domain that the request is coming from by
default, meaning if users authenticate from journaly.com then the cookie
will be scoped to "journaly.com".

The non-specified default only applies to the root domain and
doesn't propagate through to subdomains, so the cookie would not be sent
to e.g. api.journaly.com.

However, by explicitly setting the cookie domain to "journaly.com", it also
includes the subdomains:

> "The Domain attribute specifies which hosts are allowed to receive the
cookie. If unspecified, it defaults to the same host that set the
cookie, excluding subdomains. If Domain is specified, then subdomains
are always included." - [MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#Domain_attribute)